Adapt InitRTC init step to API change in the InitIts init step
With change [1] in plugins/its-base support for configuring branches
for the RTC integration during init was added. Due to this change the
InitRTC init step needs to be adapted.

In addition update the example of the plugin initialisation in the
documentation because with change [1] there is now an additional
question during init for configuring the branches.

[1] https://gerrit-review.googlesource.com/53690

Change-Id: I8e15fde9a40d3f7d64779ae17941d04307b7cc48
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>