New Feature: Optional commit validation check for allowed issue status
This feature introduces an optional field to the commitLink configuration
that can be used during commit validation to check whether the issues 
associated to the commit are in the proper state (allowed status). 
Zero to many "allowedStatus" fields can be listed under the commitLink
configuration. If, during commit validation, the issue's status does not
match that of one of the "allowedStatus" values then the commit fails and
displays a notification to the committer of the problem.

This commit covers the RTC specific portion of the its-base parent 
plugin project. Specifically it implements the getStatus interface to retrieve
the bug status from RTC.

Feature: Issue 2818
Change-Id: I8970d6358d3f5fe47e5fe9075e1e8e4dab632f75