Follow upstream's move of HTTPClient and HTTPCommons definitions
In a4640fb15cf3d5ff15cfe9028a80af1d8bf0058d, upstream moved BUCK
library definitions of HTTPClient and HTTPCore. We follow that move to
make the BUCK build pass again.

This commit only fixes the building process work again, and does not
yet address deprecation of classes.

We do not mess with the maven build, as we're about to get rid of
Maven anyways.

Change-Id: Ib840007e079a6eb04ab301b8a0bed310683d7830