Eliminate unneeded dependencies and simplify Buck rules
Replace commons-logging with slf4j and commons-io with Guava library.
Advantage is that while former are not included in Gerrit Plugin API,
the later are included.

Don't mention the libraries explicitly, because they already included
implicitly in the Plugin API.

Remove unused dependency on mockito and drop unused java_test() rule.

Change-Id: Idd0bb00ded6a838c761aaf692d2bffc9aa6336a7